feat(MB-PB-003): Wrapped api in docker container, configured interaction with DB - Fixed Domain classes so that they correspond to the class diagram - Added new Domain level classes that were missing earlier - Changed files related to database migrations - Added and customized the DockerFile needed to run the API container - Fixed problems with connecting to the database - Fixed the init.sql file, which contains the query for auto-forming and filling the database. - Fixed issues related to user and Swagger interaction when using containers

+++ Domain/Entities/Subject.cs
@@ -2,7 +2,7 @@ namespace Domain.Entities
 {
     public class Subject
     {
-        public long Id { get; set; }  // Уникальный идентификатор темы
+        public int Id { get; set; }
         public string Name { get; set; }
 
         public ICollection<BookSubject> BookSubjects { get; set; }
